requirements.txt
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
pandas
snowflake-connector-python
python-dotenv
pytube
apache-airflow